Refactor the SelectionState out of the VM
