+++ WooCommerce/src/main/kotlin/com/woocommerce/android/ui/products/selector/SelectionState.kt
@@ -1,0 +1,7 @@
+package com.woocommerce.android.ui.products.selector
+
+enum class SelectionState {
+    SELECTED,
+    UNSELECTED,
+    PARTIALLY_SELECTED
+}
